+++ postSystem/postSystem/readData/SumCalculator.cs
@@ -7,23 +7,23 @@ public class SumCalculator
     /// </summary>
     /// <param name="itemList"></param>
     /// <returns>sum</returns>
-    public static List<int> CalculateSum(List<string[]> itemList)
+    public static List<float> CalculateSum(List<string[]> itemList)
     {
-        int totalPrice = 0;
-        int totalWeight = 0;
-        int totalItems = 0;
+        float totalPrice = 0;
+        float totalWeight = 0;
+        float totalItems = 0;
 
         foreach (string[] strings in itemList)
         {
-            int packagePrice = int.Parse(strings[3]);
-            int postagePrice = int.Parse(strings[5]);
-            int weight = int.Parse(strings[1]);
+            float packagePrice = float.Parse(strings[3]);
+            float postagePrice = float.Parse(strings[5]);
+            float weight = float.Parse(strings[1]);
             totalPrice += packagePrice + postagePrice;
             totalWeight += weight;
             totalItems += 1;
         }
 
-        List<int> sum =
+        List<float> sum =
         [
             totalItems,
             totalWeight,
